+++ crates/types/src/signaling/control/event_info.rs
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
+//
+// SPDX-License-Identifier: EUPL-1.2
+
+use crate::imports::*;
+
+/// Information about an event
+#[derive(Clone, Debug, PartialEq, Eq)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+pub struct EventInfo {
+    /// The title of the event
+    pub title: String,
+}